requirements.prod.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.4
pydantic==1.10.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.17.1
slowapi==0.1.7
redis==4.5.5
structlog==22.3.0
json-log-formatter==0.5.0
gunicorn==21.2.0
